sitio de minado

+++ core/src/com/mygdx/entidades/ObjetosDelMapa/Minable/CarbonMena.java
@@ -7,7 +7,7 @@ import com.mygdx.utiles.Recursos;
 
 public class CarbonMena extends Mineral{
 
-	public CarbonMena(int x, int y, World world, boolean comprable) {
+	public CarbonMena(float x, float y, World world, boolean comprable) {
 		super(x, y, world, comprable, Recursos.MENA_CARBON, TipoMinerales.CARBON, EstadosMinerales.MENA, 10, 8);
 	}
